Handle fragmented reads on socket connection facade
Sl4a BluetoothConnectionThroughputRead reads from socket and and
evaluate the integrity of the test package expecting the entire package
to be in the buffer.  However because it is a socket stream it is not a
correct assumption that a read call always returns the entire packages
as it were sent.  This causes the command
BluetoothConnectionThroughputRead to incorrectly fail on integrity of
the package been corrupt.  The solution is to buffer the received data
until an entire buffer have been received and then match the received
data against expected package contents.

Bug: None
Change-Id: I2b83cc778f0818747672785c0d3482dc290526f2