Remove usage of StringNetworkSpecifier
StringNetworkSpecifier is a hidden API class that is removed in R, and
replaced by EthernetNetworkSpecifier and TestNetworkSpecifier.

Remove its usage in ConnectivityManagerFacade. It was misused to hold
JSON data to be later parsed and converted to a wifi NetworkSpecifier.

The current NetworkRequest builder from JSON is only used in
connectivityRegisterNetworkCallback which is unused, in
connectivityRequestNetwork which is never used with a NetworkSpecifier,
and in connectivityRequestWifiAwareNetwork. Only keep NetworkSpecifier
support in connectivityRequestWifiAwareNetwork, as the JSON format only
allows specifiers for that kind of network (generic string specifiers
should not be used as they are considered as EthernetNetworkSpecifier,
and the JSON format does not define a NetworkSpecifier type).

Bug: 179329291
Test: m sl4a
Change-Id: Iaa7a5395cd6f2af7f788b25276d8ba6c4c0eb2a2