Remove usage of bluetooth hidden api
Most of Sl4a x Bluetooth usage are not in a working shape. But because
the code still exist and link to the internal implementation, it prevent
us from making internal update.

Removing the access to the hidden api can only be done with theses
changes.

This CL try to keep most of the api working, and even fix some of the
non working one to the correct one. But for api that are based on
internal data like and not intented to be exposed like the gatt instance
id or like managing the phone connection status, there is no other
option than deleting them

Test: m sl4a.Common
Bug: 450679763
Flag: EXEMPT TEST_ONLY
Change-Id: Id473993002eeefc4372968360c4d05eb33b64fa8